update general layout
