+++ client/src/components/Layout/index.js
@@ -4,7 +4,6 @@ import PropTypes from 'prop-types';
 import General from './General';
 import SplitScreen from './SplitScreen';
 import Dashboard from './Dashboard';
-import StepLayout from './Step';
 
 const Layout = ({ layout, ...props }) => {
   switch (layout) {
@@ -12,8 +11,6 @@ const Layout = ({ layout, ...props }) => {
       return <Dashboard {...props} />;
     case 'splitScreen':
       return <SplitScreen {...props} />;
-    case 'step':
-      return <StepLayout {...props} />;
     case 'general':
     default:
       return <General {...props} />;
