+++ client/src/pages/Section/style.js
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+import { Link as RLink } from 'react-router-dom';
+import setMargin from '../../helpers/set-margin';
+
+export const Container = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+`;
