+++ client/src/pages/index.js
@@ -1,9 +1,9 @@
 import Home from './Home';
-import Step from './Step';
+import Section from './Section';
 import Login from './Login';
 import Signup from './Signup';
 import ForgotPassword from './ForgotPassword';
 import ResetPassword from './ResetPassword';
 import Admin from './Admin';
 
-export { Admin, Home, Step, Login, Signup, ForgotPassword, ResetPassword };
+export { Admin, Home, Section, Login, Signup, ForgotPassword, ResetPassword };
